ChargingPlatform/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
ocpp>=0.19.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
python-dateutil>=2.8.2
pymysql>=1.1.0
cryptography>=41.0.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0